implement generic vectorized algorithm to unpack 32 bits integers
